Makes translations work better
- Changes \if's to be positive

The options were negated, which easily leads to errors (which it did).
Even if the option is `nobeamer` the code should be
`\if@didactic@beamer` instead of `\if@didactic@nobeamer`.

- Changes to use `translations` instead of `translator`

This way we can use the normal `babel` way and not have to provide the
languages as options to the class.

- Adds use of `\crefname` and `\Crefname` to fix bug

Without the use of `\crefname` and `\Crefname` the translations don't
work for `refname` and `Refname` in `thmtools`.