Makes starred versions use full width
This is a breaking change, as the default is to not use the full width
(as it was before).

We add `\textbytext*`, `\codebycode*`, and `\runbyrun*` that use the
full width and blocks margin. This means that the unstarred versions are
safe to use in floats, while the starred versions are not.